refactor: move DmChannelType to channel.type file
