+++ client/src/component/organism/DmCard/index.ts
@@ -1,17 +1,5 @@
 /* eslint-disable no-unused-vars */
-import { UserType } from '@type/user.type'
-
-export interface DmChannelType {
-  id: number
-  name: string
-  type: 'PRIVATE' | 'PUBLIC' | 'DM'
-  createdAt?: string
-  updatedAt?: string
-  deletedAt?: string
-  workspaceId: number
-  memberCount?: number
-  memberMax3: UserType[]
-}
+import { DmChannelType } from '@type/channel.type'
 
 export interface DmCardProps {
   dmChannel: DmChannelType
