+++ client/src/type/channel.type.ts
@@ -9,6 +9,18 @@ export interface ChannelCardType extends Object {
   joined: boolean
 }
 
+export interface DmChannelType {
+  id: number
+  name: string
+  type: 'PRIVATE' | 'PUBLIC' | 'DM'
+  createdAt?: string
+  updatedAt?: string
+  deletedAt?: string
+  workspaceId: number
+  memberCount?: number
+  memberMax3: UserType[]
+}
+
 export interface ChannelType {
   id: number
   name: string
